+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@
 
 #### Items
 - Changed Quiver use action to equip it, rather than empty its contents
+- Changed Dagger to be able to parry swords in blade clashes
 
 #### Blocks
 - Changed Iron Grates to have the `dragon_immune` block tag